+++ src/main/kotlin/dev/aaronhowser/mods/excessive_utilities/effect/OilyEffect.kt
@@ -1,10 +1,13 @@
 package dev.aaronhowser.mods.excessive_utilities.effect
 
+import dev.aaronhowser.mods.excessive_utilities.config.ServerConfig
+import net.minecraft.tags.DamageTypeTags
 import net.minecraft.world.effect.MobEffect
 import net.minecraft.world.effect.MobEffectCategory
 import net.minecraft.world.entity.LivingEntity
 import net.minecraft.world.entity.player.Player
 import net.minecraft.world.phys.Vec3
+import net.neoforged.neoforge.event.entity.living.LivingIncomingDamageEvent
 
 class OilyEffect : MobEffect(
 	MobEffectCategory.HARMFUL,
@@ -29,6 +32,15 @@ class OilyEffect : MobEffect(
 		}
 
 		return true
+	}
+
+	companion object {
+		fun handleIncomingDamage(event: LivingIncomingDamageEvent) {
+			if (event.isCanceled) return
+			if (!event.source.`is`(DamageTypeTags.IS_FIRE)) return
+
+			event.amount *= ServerConfig.CONFIG.oilyFireDamageFactor.get().toFloat()
+		}
 	}
 
 }